PubSub messages are hopefully here
The implementation of pubsub module is practically there
PubSub messages are now received via MPI and unpacked.
PubSub messages are processed following a path parallel to that of normal messages
They should work nonetheless. Had to remove tests to make the build work.
It compiles and then rootsim-cc compiled a model with a stub of the functions needed to work.

This still needs to be tested for correctness.

TODO:
The messages still do not get garbage collected from the past array.

+++ src/lp/lp.h
@@ -30,7 +30,7 @@ struct lp_ctx {
 	/// Bitmap for subbed nodes
 	struct block_bitmap* subnodes;
     /// Count of nodes subbed to this LP
-    size_t n_remote_sub_nodes;
+    int n_remote_sub_nodes;
 #endif
 
 	/// The memory allocator state of this LP
